Weather calls Alpaca API
Before the controller was simply reading a text file, but now it has been modified to call the Alpaca API on an observatory's weather sensors.  It makes several concurrent calls for each weather metric and aggregates them into a weather status object.  Also included are a set of tests that mock the Alpaca call.