Miscellaneous fixes
I had failed to update the readme with the new way observe-weather works.  Also, there were some syntax errors that were fixed as well as a problem with the weather status object.  It's properties are being used to contruct the API URLs, but under this approach some of the properties had the wrong names.  The proper corrections have been made.