Pull request  comments and cleaning up
Some of the code has been cleaned up to make it less convoluted and pull request comments have been addressed.  In particular the axios.all call will now behave even if an individual call fails.  This is due to the indivdual call function returning a null value when it fails.