Added precision and fixed security
Due to the precise data returned by Alpaca, sometimes the client ends up updating weather data often when there are only small changes to the data.  In order to reduce the amount of updates, this new code takes in a 'precision' parameter to the config object that tells the controller how many significant digits to return to the client.  Then the controller returns the data formatted that way.  Also included are some dependency changes which should fix some security problems that GitHub showed on the repository page.