Another try for security
My last attempt at fixing the security issues that GitHub notified us of didn't work.  I think this is because when I used 'yarn upgrade', it didn't remove the old version of lodash from the yarn.lock file for some reason.  So I went in and and changed every reference of the old lodash to the new lodash.  Hopefully this fixes the security problems.